Added an sst.merlin.interfaces module and moved LinkControl and ReorderLinkControl to it.
