+++ src/sst/elements/merlin/merlin.cc
@@ -51,6 +51,10 @@ char pymerlin_endpoint[] = {
 #include "pymerlin-endpoint.inc"
     0x00};
 
+char pymerlin_interfaces[] = {
+#include "interfaces/pymerlin-interfaces.inc"
+    0x00};
+
 char pymerlin_topo_dragonfly[] = {
 #include "topology/pymerlin-topo-dragonfly.inc"
     0x00};
@@ -75,6 +79,7 @@ public:
         auto primary_module = createPrimaryModule(pymerlin,"pymerlin.py");
         primary_module->addSubModule("base",pymerlin_base,"pymerlin-base.py");
         primary_module->addSubModule("endpoint",pymerlin_endpoint,"pymerlin-endpoint.py");
+        primary_module->addSubModule("interfaces",pymerlin_interfaces,"interfaces/pymerlin-interfaces.py");
         primary_module->addSubModule("topology",pymerlin_topo_dragonfly,"topology/pymerlin-topo-dragonfly.py");
         primary_module->addSubModule("topology",pymerlin_topo_hyperx,"topology/pymerlin-topo-hyperx.py");
         primary_module->addSubModule("topology",pymerlin_topo_fattree,"topology/pymerlin-topo-fattree.py");
